Updated tests for return to home to include dials and touch
