+++ tests/test_app_plus.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import asyncio
 import json
 from pathlib import Path
 from typing import Any
@@ -21,6 +22,7 @@ from home_assistant_streamdeck_yaml import (
     Page,
     TouchscreenEventType,
     _on_dial_event_callback,
+    _on_press_callback,
     _on_touchscreen_event_callback,
     _update_state,
     update_dial,
@@ -396,3 +398,127 @@ async def test_touchscreen(
     )
     attributes = dial.get_attributes()
     assert attributes["state"] is not attributes["min"]
+
+
+async def test_return_to_home(  # noqa: PLR0915
+    mock_deck_plus: Mock,
+    websocket_mock: Mock,
+    state: dict[str, dict[str, Any]],
+    dials: list[Dial],
+) -> None:
+    """Test that the return to home works."""
+    return_to_home_after = 0.8
+    assert return_to_home_after
+    shorter_than_return_to_home_after = return_to_home_after - 0.1
+    assert shorter_than_return_to_home_after < return_to_home_after
+    assert shorter_than_return_to_home_after > 0.0
+    longer_than_return_to_home_after = return_to_home_after + 0.1
+    assert longer_than_return_to_home_after > return_to_home_after
+    assert longer_than_return_to_home_after > 0.0
+    longer_and_shorter_delta = (
+        longer_than_return_to_home_after - shorter_than_return_to_home_after
+    )
+    home = Page(
+        name="home",
+        buttons=[
+            Button(special_type="go-to-page", special_type_data="anon"),
+            Button(special_type="go-to-page", special_type_data="second"),
+        ],
+    )
+    second = Page(
+        name="second",
+        buttons=[Button(special_type="empty")],
+        dials=dials,
+    )
+    anon = Page(
+        name="anon",
+        buttons=[Button(text="yolo"), Button(text="foo", delay=0.1)],
+    )
+    config = Config(
+        pages=[home, second],
+        anonymous_pages=[anon],
+        return_to_home_after_no_presses={
+            "home_page": "home",
+            "duration": return_to_home_after,
+        },
+    )
+    inactivity_state = InactivityState()
+
+    # We need to have a release otherwise it will be timing for a long press
+    async def press_and_release(key: int) -> None:
+        press = _on_press_callback(inactivity_state, websocket_mock, state, config)
+        await press(mock_deck_plus, key, True)  # noqa: FBT003
+        await press(mock_deck_plus, key, False)  # noqa: FBT003
+
+    assert config._current_page_index == 0
+    assert config._detached_page is None
+    await press_and_release(0)
+    assert config._detached_page is not None
+    assert config.current_page() == anon
+    await asyncio.sleep(
+        longer_than_return_to_home_after,
+    )  # longer than delay should then switch to home
+    # Should now be on the home page, with the anon page closed automatically
+    assert config._detached_page is None
+    assert config.current_page() == home
+    # Check that the logic also works on non detached pages
+    await press_and_release(1)
+    assert config._detached_page is None
+    assert config.current_page() == second
+    await asyncio.sleep(
+        shorter_than_return_to_home_after,
+    )  # shorter than delay should stay on page
+    assert config._detached_page is None
+    assert config.current_page() == second
+    await asyncio.sleep(longer_and_shorter_delta)
+    assert config._detached_page is None
+    assert config.current_page() == home
+    # Check that multiple button presses delay the return to home
+    # We have to use a non-detached page for these, as these close
+    # When pressed
+    await press_and_release(1)
+    assert config._detached_page is None
+    assert config.current_page() == second
+    await asyncio.sleep(
+        shorter_than_return_to_home_after,
+    )  # shorter than delay should stay on page
+    assert config._detached_page is None
+    assert config.current_page() == second
+    await press_and_release(0)
+    await asyncio.sleep(
+        shorter_than_return_to_home_after,
+    )  # shorter than delay, should still remain on page
+    assert config._detached_page is None
+    assert config.current_page() == second
+
+    # Check that dial events also reset the timer
+    async def dial_something() -> None:
+        dial = _on_dial_event_callback(inactivity_state, websocket_mock, state, config)
+        await dial(mock_deck_plus, 0, DialEventType.TURN, 1)
+
+    await dial_something()
+    await asyncio.sleep(
+        shorter_than_return_to_home_after,
+    )  # shorter than delay, should still remain on page
+    assert config._detached_page is None
+    assert config.current_page() == second
+
+    # Check that dial events also reset the timer
+    async def touch_something() -> None:
+        touch = _on_touchscreen_event_callback(
+            inactivity_state,
+            websocket_mock,
+            state,
+            config,
+        )
+        await touch(mock_deck_plus, TouchscreenEventType.SHORT, {"x": 0, "y": 0})
+
+    await touch_something()
+    await asyncio.sleep(
+        shorter_than_return_to_home_after,
+    )  # shorter than delay, should still remain on page
+    assert config._detached_page is None
+    assert config.current_page() == second
+
+    await asyncio.sleep(longer_than_return_to_home_after)
+    assert config.current_page() == home
